[Bronze V] Title: 이상한 기호, Time: 176 ms, Memory: 17684 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -3,13 +3,13 @@ import java.util.Scanner;
 public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        int A = sc.nextInt();
-        int B = sc.nextInt();
+        long A = sc.nextInt();
+        long B = sc.nextInt();
 
         System.out.println(addNum(A,B));
     }
 
-    static int addNum(int num1, int num2) {
+    static long addNum(long num1, long num2) {
         return (num1+num2)*(num1-num2);
     }
 }
